Added delete for Account type

+++ backend/routes/accountTypes.ts
@@ -3,6 +3,7 @@ import mongoose from "mongoose";
 import auth from "../middleware/auth";
 import AccountType from "../models/AccountType";
 import {imagesUpload} from "../multer";
+import Account from "../models/Account";
 
 const accountTypesRouter = express.Router();
 
@@ -51,6 +52,15 @@ accountTypesRouter.post('/', auth, imagesUpload.single('image'), async (req, res
 
 accountTypesRouter.delete('/:id', auth, async (req, res, next) => {
   try {
+    const accountType = await AccountType.findById(req.params.id);
+    console.log(accountType)
+    const account = await Account.find({accountType: accountType?._id})
+    console.log(account)
+
+    if (account.length) {
+      return res.status(403).send({message: 'First delete all accounts with this type'})
+    }
+
     const result = await AccountType.deleteOne({_id: req.params.id});
 
     if (result.deletedCount) {
